Incorporate previous changes made for debian
These set of changes had been applied previously. The changes made
include setting the mail spool location to comply with Debian policy;
change the lib localtion from /usr/locqal/lib to /usre/lib/, set the
location of the documentation to be /usr/share/doc, and not in
/usr/lib, and mentioning that mailagent is not, as required of
programs that look at the spool directory sgid mail (for locking),
since that would be a security risk.  This requires a final default
rule to always save the mail locally.  The C filter has also been
updated to workj with changes in recent compilers.

git-archimport-id: srivasta@debian.org--2003-primary/mailagent--debian--3.0--patch-1