Fix mailagent to correctly parse top level domains
The following patched code in mailagent examines strings in Received:
headers and makes three decisions based on whether the string appears
to be a FQDN.  The first two decisions help Mailagent construct an
accurate list of hosts through which the mail was relayed.  The last
decision logs a warning message if the hostname string that Mailagent
derives seems to be "bad."

Top level domains with more than four characters are "bad" for
Mailagent, but there are two official TLDs with more than four
characters: ".museum" and ".travel" (see
http://www.iana.org/gtld/gtld.htm).  It's also useful to use the TLD
".private" for private IP networks.

Mailagent rules can branch on the list of relay hosts; such rules may
not function as expected when a long TLD appears in a Recieved:
header.  I consider this consequence to have a severity of "normal";
the other consequences are incorrect log entries and spurious warnings
in the log, which have a severity of "minor."

* agent/pl/parse.pl (relay_list): Added code which compares the TLD with
  the list of official TLDs /lus ".private".  The patch eliminates false
  negatives for the .museum, and .travel TLDs, allows the TLD .private,
  and eliminates false positives for all strings of length 2-4 that are
  not official TLDs. 


git-archimport-id: srivasta@debian.org--etch/mailagent--debian--3.0--patch-3