[master]:  Messages should be dropped in the system mailbo
When the LEAVE command is used, messages should be dropped into the
system mailbox of the user (/var/spool/mail/$user, which is the value
of the $mailfile variable, defined in /usr/bin/mailagent ).  Instead
of that, the messages are dropped into ~/mbox.$user.  The patch
attached to this bug report fixes the problem.  The bug comes clearly
from a typo and it is surprising that it has not yet been noticed.

Bug fix: "Messages should be dropped in the system mailbox when using
the LEAVE command", thanks to Rafael Laboissiere (Closes: #812308).

Signed-off-by: Manoj Srivastava <srivasta@debian.org>